chore: treat all warnings as errors except one rule: dead-code (#46)
This commit try to fix various code/lint issues: unused import, non snake_case variables,
unmatched pattern arms etc. After the fix, we can enable treating all warnings as errors.

The purpose of this commit is to reduce the warning logs when building, which is annoying
to find the real errors/problems.